examples/asp/dwarf/requirements.txt
pyelftools
pytest
pyaml
